--- alchemist-adapter/build.gradle.kts
@@ -1,5 +1,0 @@
-import Libs.alchemist
-
-dependencies {
-    implementation(alchemist("loading")) // to use the custom exporter for multivesta
-}
