--- alchemist-adapter/src/main/kotlin/it/unibo/alchemist/adapter/SimultationAdapter.java
@@ -1,18 +1,0 @@
-/*
- * Copyright (C) 2010-2023, Danilo Pianini and contributors
- * listed, for each module, in the respective subproject's build.gradle.kts file.
- *
- * This file is part of Alchemist, and is distributed under the terms of the
- * GNU General Public License, with a linking exception,
- * as described in the file LICENSE in the Alchemist distribution's top directory.
- */
-
-package it.unibo.alchemist.adapter;
-
-public interface SimultationAdapter {
-    double getTime();
-    double rval(String obs);
-    double rval(int obsId);
-    void doStep();
-    void performWholeSimulation();
-}
