+++ alchemist-multivesta-adapter/src/main/kotlin/it/unibo/alchemist/multivesta/adapter/multivesta/MultiVestaEntryPoint.kt
@@ -1,0 +1,19 @@
+/*
+ * Copyright (C) 2010-2023, Danilo Pianini and contributors
+ * listed, for each module, in the respective subproject's build.gradle.kts file.
+ *
+ * This file is part of Alchemist, and is distributed under the terms of the
+ * GNU General Public License, with a linking exception,
+ * as described in the file LICENSE in the Alchemist distribution's top directory.
+ */
+
+package it.unibo.alchemist.multivesta.adapter.multivesta
+
+import entrypointmultivesta.UniqueEntryPoint
+
+object MultiVestaEntryPoint {
+    @JvmStatic
+    fun main(args: Array<String>) {
+        UniqueEntryPoint.main(args)
+    }
+}
