+++ build.gradle.kts
@@ -153,7 +153,7 @@ allprojects {
 
     tasks.withType<KotlinCompile>().configureEach {
         kotlinOptions {
-            jvmTarget = "1.8"
+            jvmTarget = "11"
             freeCompilerArgs = listOf(
                 "-Xjvm-default=all", // Enable default methods in Kt interfaces
                 // Context receivers temporarily disabled, as they are unsupported in Kotlin script
@@ -322,7 +322,6 @@ dependencies {
     implementation(libs.apache.commons.cli)
     implementation(libs.guava)
     implementation(libs.logback)
-    implementation(alchemist("adapter"))
     testRuntimeOnly(incarnation("protelis"))
     testRuntimeOnly(incarnation("sapere"))
     testRuntimeOnly(incarnation("biochemistry"))
