--- multivesta/.gitignore
@@ -1,1 +1,0 @@
-MultiVeStA_OUTPUT/
