--- multivesta/README.md
@@ -1,3 +1,0 @@
-# multialchemista
-
-Integration of [MultiVeStA](https://github.com/andrea-vandin/MultiVeStA) in [Alchemist](https://github.com/AlchemistSimulator/Alchemist).
