--- multivesta/src/main/java/samplemvintegration/multivesta/EntryPointSampleIntegration.java
@@ -1,10 +1,0 @@
-package samplemvintegration.multivesta;
-
-import entrypointmultivesta.UniqueEntryPoint;
-
-public class EntryPointSampleIntegration {
-
-	public static void main(String[] args) {
-		UniqueEntryPoint.main(args);
-	}
-}
